# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/receipts/process": {
            "post": {
                "description": "Submits a receipt and returns a unique ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Receipts"
                ],
                "summary": "Process a receipt and generate an ID",
                "parameters": [
                    {
                        "description": "Receipt JSON",
                        "name": "receipt",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Receipt"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/receipts/{id}/points": {
            "get": {
                "description": "Returns points for a given receipt ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Receipts"
                ],
                "summary": "Get points for a receipt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Receipt ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Item": {
            "description": "Item model for the receipt",
            "type": "object",
            "properties": {
                "price": {
                    "type": "string",
                    "example": "6.49"
                },
                "shortDescription": {
                    "type": "string",
                    "example": "Mountain Dew 12PK"
                }
            }
        },
        "models.Receipt": {
            "description": "Receipt model for processing",
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Item"
                    }
                },
                "purchaseDate": {
                    "type": "string",
                    "example": "2022-01-01"
                },
                "purchaseTime": {
                    "type": "string",
                    "example": "13:01"
                },
                "retailer": {
                    "type": "string",
                    "example": "Target"
                },
                "total": {
                    "type": "string",
                    "example": "35.35"
                }
            }
        }
    }
}